+++ src/view/add-form.js
@@ -1,4 +1,4 @@
-import AbstractView from './Abstract';
+import AbstractView from './abstract';
 
 const createAddFormTemplate = () => (
   '<p class="trip-events__msg">Click New Event to create your first point</p>'
